refactor: relocate isDirectory and isFile
